+++ crates/common/src/scripts/functions/array.rs
@@ -26,9 +26,9 @@ pub fn fn_sort<'x>(_: &'x Context<'x>, v: Vec<Variable>) -> Variable {
     let is_asc = v[1].to_bool();
     let mut arr = (*v[0].to_array()).clone();
     if is_asc {
-        arr.sort_unstable_by(|a, b| b.cmp(a));
-    } else {
         arr.sort_unstable();
+    } else {
+        arr.sort_unstable_by(|a, b| b.cmp(a));
     }
     arr.into()
 }